test setup works now + minor config changes

+++ setup.sh
@@ -4,6 +4,10 @@
 # Exit immediately if a command exits with a non-zero status
 set -e
 
+SCRIPT_DIR=$( cd -- "$( dirname -- "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )
+
+cd "$SCRIPT_DIR"
+
 # Parse command-line arguments
 NO_GITHUB=false
 NO_UPDATE=false
@@ -30,11 +34,21 @@ if [ "$NO_UPDATE" = false ]; then
 fi
 
 # Install necessary packages
-sudo apt install gh neovim tmux nodejs gcc xclip ripgrep python3-venv xdg-utils -y
+sudo chmod +x install_deps.sh
+bash install_deps.sh
 
 # Install NVM and Node.js
-curl -o- https://raw.githubusercontent.com/nvm-sh/nvm/v0.40.2/install.sh | bash
-. "$HOME/.nvm/nvm.sh" && nvm install node
+if [ ! -s "$HOME/.nvm/nvm.sh" ]; then
+    echo "Installing nvm..."
+    curl -o- https://raw.githubusercontent.com/nvm-sh/nvm/v0.40.2/install.sh | bash
+fi
+
+# Load nvm and check if node exists
+. "$HOME/.nvm/nvm.sh"
+if ! command -v node &> /dev/null; then
+    echo "Installing node..."
+    nvm install node
+fi
 
 # Authenticate with GitHub CLI if --no-github flag is not set
 if [ "$NO_GITHUB" = false ]; then
@@ -67,9 +81,24 @@ sudo apt install -y python3-libtmux
 git clone https://github.com/ofirgall/tmux-window-name \
     ~/.config/tmux/plugins/tmux-window-name
 
+# Install neovim if not already installed
+if ! command -v nvim &> /dev/null; then
+    echo "Installing Neovim..."
 
-# Inject into .bashrc
-echo "source ~/.config/bash/bash_config.sh" >> ~/.bashrc
+    # try to install Neovim using snap if not built from source
+    if command -v snap &> /dev/null; then
+        sudo snap install nvim --classic
+    else
+        # If snap is not available, try to build from source
+        echo "Snap not found, building Neovim from source..."
+        git clone https://github.com/neovim/neovim /usr/local/src/neovim
+        cd /usr/local/src/neovim
+        make CMAKE_BUILD_TYPE=RelWithDebInfo
+        sudo make install
+        cd "$SCRIPT_DIR"
+    fi
+fi
+
 
 
 # Install packer.nvim
@@ -101,33 +130,20 @@ if [ "$NO_GITHUB" = false ]; then
     nvim -c "Copilot setup" -c "q!"
 fi
 
-repo_dir=$(pwd)
 
-# Find all files and directories in TARGET_CONFIG
-echo Creating link for config files...
-find ".config" | while read -r path; do
-    # Determine relative path
-    source_path="$repo_dir/$path"
-    target_path="$HOME/$path"
+if [ "$SYM_LINK" = true ]; then
+    bash link.sh --symlink
+else
+    bash link.sh
+fi
 
-    if [ -d "$source_path" ]; then
-        mkdir -p "$target_path"
-    elif [ -f "$source_path" ]; then
-        if [ "$SYM_LINK" = true ]; then
-            ln -sv "$source_path" "$target_path"
-        else
-            ln -sfv "$source_path" "$target_path"
-        fi
-    fi
-done
 
+# Inject into .bashrc
+echo "source ~/.config/bash/bash_config.sh" >> ~/.bashrc
 
 # Grant execute permission to all scripts
 find .config -type f -name "*.sh" -o -name "*.py" | while read -r path; do
     sudo chmod +x "$path"
 done
 
-
-
 echo "SETUP COMPLETE! run 'exec bash' to load new config"
-
